reorganize files with role-specific naming convention
