+++ client/src/screens/FertilizerAdvisor/NLPAdvisoryFarmer/NLPAdvisoryFarmerResultsScreen.tsx
